requirements.txt
flask
google-generativeai
Pillow
SpeechRecognition
pyttsx3
gunicorn